+++ app/src/main/java/com/byeboo/app/presentation/quest/review/common/all/CommonAnswerScreen.kt
@@ -1,4 +1,4 @@
-package com.byeboo.app.presentation.quest.review.common
+package com.byeboo.app.presentation.quest.review.common.all
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
@@ -19,7 +19,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.vectorResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
